Move NameValuePermission to be a subclass of SqlClientPermission (it's only used there)
